+++ packages/ganache-deployer/.prettierignore
@@ -1,0 +1,3 @@
+contracts/AssetHolder.sol
+contracts/NitroAdjudicator.sol
+contracts/examples/SingleAssetPayments.sol
