+++ packages/ganache-deployer/.prettierrc.js
@@ -1,0 +1,13 @@
+module.exports = {
+  ...require('../../.prettierrc.js'),
+  "trailingComma": "es5",
+  "printWidth": 100,
+  "overrides": [
+    {
+      "files": "*.sol",
+      "options": {
+        "tabWidth": 4
+      }
+    }
+  ]
+};
