+++ packages/ganache-deployer/tslint.json
@@ -1,0 +1,4 @@
+{
+  "extends": "../../tslint.json",
+  "exclude": ["node_modules/**", "lib/**"]
+}
